requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
bcrypt==4.1.2
SQLAlchemy==2.0.23
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-Migrate==4.0.5
python-dotenv==1.0.0
Flask-WTF==1.2.1
WTForms==3.1.1
Flask-Admin==1.6.3
Flask-RESTful==0.3.10
